fix: explicitly cast seconds duration to uint256 to avoid overflows

+++ contracts/ERC20StakingRewardsDistribution.sol
@@ -178,7 +178,7 @@ contract ERC20StakingRewardsDistribution {
             // recoverable rewards are going to be recovered in this tx (if it does not revert),
             // so we add them to the claimed rewards right now
             _reward.claimed += ((_reward.recoverableSeconds * _reward.amount) /
-                (secondsDuration * MULTIPLIER));
+                (uint256(secondsDuration) * MULTIPLIER));
             delete _reward.recoverableSeconds;
             uint256 _recoverableRewards =
                 IERC20(_reward.token).balanceOf(address(this)) -
@@ -406,7 +406,7 @@ contract ERC20StakingRewardsDistribution {
                 uint256 _nonRequiredFunds =
                     _reward.claimed +
                         ((_reward.recoverableSeconds * _reward.amount) /
-                            (secondsDuration * MULTIPLIER));
+                            (uint256(secondsDuration) * MULTIPLIER));
                 return
                     IERC20(_reward.token).balanceOf(address(this)) -
                     (_reward.amount - _nonRequiredFunds);
